Separate constants and utility functions
