+++ geometa/constants.py
@@ -1,0 +1,1 @@
+GEOSPATIAL_FIELD = 'geometa.bounds'
