+++ geometa/exceptions.py
@@ -1,0 +1,2 @@
+class CannotHandleError(Exception):
+    pass
