+++ types/grid/geometa_grid/schema.py
@@ -1,6 +1,7 @@
 from marshmallow import fields
 from geometa.schema import BaseSchema
-from geometa import from_bounds_to_geojson, CannotHandleError
+from geometa.utils import from_bounds_to_geojson
+from geometa.exceptions import CannotHandleError
 import gdal
 import osr
 from shapely.geometry import Polygon
