+++ types/raster/geometa_raster/schema.py
@@ -1,6 +1,7 @@
 from marshmallow import fields, ValidationError
 from geometa.schema import BaseSchema
-from geometa import from_bounds_to_geojson, CannotHandleError
+from geometa.utils import from_bounds_to_geojson
+from geometa.exceptions import CannotHandleError
 import gdal
 import osr
 import rasterio
